Extract LogCache.Poster task
